Removed testing force click data
