+++ tests/client-test/redux-test/actions-test/results-actions.spec.js
@@ -1,0 +1,9 @@
+/*  eslint-env mocha */
+import { expect } from 'chai';
+import * as action from '../../../../client/redux/actionCreators/modal-actions';
+
+describe('Redux Results Actions', () => {
+  describe('', () => {
+    
+  };
+});
